deploy/docker/requirements.txt
joblib==0.16.0
numpy==1.19.1
pandas==1.1.0
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
vtreat==0.4.6
PyYAML==5.3.1
lightgbm==3.0.0
xgboost==1.2.0